finished program
Still a bit of work to be done.
I need to:

find a way to move the python script to another CPU to reduce latency.

Build a script for dropbox. The need for authentication has the potential to mess up the automated process if the credentials expire. Murphy's Law.

clean up the python code, It looks ugly. it's also sending the files with full paths.

scale down the necessary size for motion detection. may get more false positives but this camera will probably be pointing down a hallway.

but these things aside, it works. I'll probably put it to use in a couple weeks.